Change orderGlobal to Float

+++ ebl/chronology/chronology.py
@@ -28,9 +28,9 @@ class Chronology:
         except IndexError:
             return None
 
-
+ 
 class KingSchema(Schema):
-    order_global = fields.Integer(data_key="orderGlobal")
+    order_global = fields.Float(data_key="orderGlobal")
     group_with = fields.Integer(
         data_key="groupWith", allow_none=True, load_default=None
     )
